pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nessusanalyzer"
version = "0.1.0"
description = "A modern Nessus vulnerability output parser and reporting toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["nessus", "vulnerability", "security", "scanner", "parser", "reporting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "click>=8.1.0",
    "lxml>=5.1.0",
    "pydantic>=2.5.0",
    "openpyxl>=3.1.0",
    "fpdf2>=2.7.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
nessusanalyzer = "nessusanalyzer.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nessusanalyzer*"]
